MAZeroingWeakRef: Fix compile error after pod update
The `MAZeroingWeakRef` pod started causing problems with more recent versions of CocoaPods. It was the same symptom as outlined in https://github.com/CocoaPods/CocoaPods/issues/4952
> clang: error: -fobjc-arc is not supported on versions of OS X prior to 10.6

This PR updates the OSX deployment target to 10.6 and removes ARC. Both changes are needed for it to compile - I have tested with either/or and it was insufficient.